+++ webfinger/__init__.py
@@ -12,8 +12,7 @@ In this module, the following are exposed:
     - BaseWebFingerClient (from webfinger.client)
     - WebFingerClient (from webfinger.client.requests  for backwards
       compatibility)
-    - The WebFingerResponse and WebFingerBuilder objects (from
-      webfinger.objects)
+    - The WebFingerJRD object (from webfinger.objects.jrd)
     - Exceptions (from webfinger.exceptions)
     - A simple helper for basic finger requests (the finger function)
 """
@@ -25,8 +24,7 @@ __version__ = "3.0.0.dev1"
 from webfinger.client import BaseWebFingerClient
 from webfinger.client.requests import WebFingerClient
 from webfinger.objects import RELS, REL_NAMES
-from webfinger.objects.response import WebFingerResponse
-from webfinger.objects.builder import WebFingerBuilder
+from webfinger.objects.jrd import WebFingerJRD
 from webfinger.exceptions import *