+++ webfinger/client/aiohttp.py
@@ -13,7 +13,6 @@ import logging
 import aiohttp
 
 from webfinger.client import BaseWebFingerClient
-from webfinger.objects.response import WebFingerResponse
 from webfinger.exceptions import WebFingerHTTPError, WebFingerNetworkError, \
     WebFingerJRDError, WebFingerContentError