+++ webfinger/client/requests.py
@@ -12,7 +12,6 @@ import requests
 import logging
 
 from webfinger.client import BaseWebFingerClient
-from webfinger.objects.response import WebFingerResponse
 from webfinger.exceptions import WebFingerHTTPError, WebFingerNetworkError, \
     WebFingerJRDError, WebFingerContentError